Add function to postprocess file directory
After archiving is complete, `postprocessing()` automatically
reorganizes the file directory to allow for .html pages in archive_html\
to retrieve html and css assets needed for re-rendering. pywebcopy
archives require a specific file structure for re-rendering to occur.
The file structure has to mimic what is found in each pywebcopy archive
folder with some assets one folder above and some in the same folder as
the .html